+++ script/tool/lib/src/main.dart
@@ -19,13 +19,13 @@ import 'java_test_command.dart';
 import 'license_check_command.dart';
 import 'lint_podspecs_command.dart';
 import 'list_command.dart';
+import 'native_test_command.dart';
 import 'publish_check_command.dart';
 import 'publish_plugin_command.dart';
 import 'pubspec_check_command.dart';
 import 'test_command.dart';
 import 'version_check_command.dart';
 import 'xcode_analyze_command.dart';
-import 'xctest_command.dart';
 
 void main(List<String> args) {
   const FileSystem fileSystem = LocalFileSystem();
