+++ script/tool/lib/src/native_test_command.dart
