+++ script/tool/test/native_test_command_test.dart
@@ -10,7 +10,7 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
-import 'package:flutter_plugin_tools/src/xctest_command.dart';
+import 'package:flutter_plugin_tools/src/native_test_command.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
